Retrieve Replies

Retrieve replies to individual conversations on a Slack team.

replies(
  ts,
  channel,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf,
  limit = 1000L
)

conversation_replies(
  conversation,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf,
  limit = 1000L
)

Arguments

ts

Character. Unique identifier of a thread's parent message.

channel

Character. Channel name to retrieve conversation from.

...

Arguments to pass to method

token

Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")

max_results

Integer. The maximum number of results to return (total). Note: The actual maximum will be max_results rounded up to a multiple of limit. Default: Inf

max_calls

Integer. The maximum number of separate API calls to make while constructing the response. Default: Inf

limit

Numeric, maximum number of results per call. Default: 1000L

conversation

An individual conversation object from the list returned by conversations.